construct audit logger registry in bootstrap builder
