+++ src/core/xds/grpc/xds_audit_logger_registry.cc
@@ -16,12 +16,11 @@
 
 #include "src/core/xds/grpc/xds_audit_logger_registry.h"
 
-#include <grpc/support/port_platform.h>
-
 #include <optional>
 #include <string>
 #include <utility>
 
+// FIXME: audit
 #include "envoy/config/core/v3/extension.upb.h"
 #include "envoy/config/rbac/v3/rbac.upb.h"
 #include "src/core/lib/security/authorization/audit_logging.h"
@@ -35,37 +34,7 @@
 
 namespace grpc_core {
 
-namespace {
-
 using experimental::AuditLoggerRegistry;
-
-class StdoutLoggerConfigFactory final
-    : public XdsAuditLoggerRegistry::ConfigFactory {
- public:
-  Json::Object ConvertXdsAuditLoggerConfig(
-      const XdsResourceType::DecodeContext& /*context*/,
-      absl::string_view /*configuration*/,
-      ValidationErrors* /*errors*/) override {
-    // Stdout logger has no configuration right now. So we don't process the
-    // config protobuf.
-    return {};
-  }
-
-  absl::string_view type() override { return Type(); }
-  absl::string_view name() override { return "stdout_logger"; }
-
-  static absl::string_view Type() {
-    return "envoy.extensions.rbac.audit_loggers.stream.v3.StdoutAuditLog";
-  }
-};
-
-}  // namespace
-
-XdsAuditLoggerRegistry::XdsAuditLoggerRegistry() {
-  audit_logger_config_factories_.emplace(
-      StdoutLoggerConfigFactory::Type(),
-      std::make_unique<StdoutLoggerConfigFactory>());
-}
 
 Json XdsAuditLoggerRegistry::ConvertXdsAuditLoggerConfig(
     const XdsResourceType::DecodeContext& context,
@@ -120,4 +89,5 @@ Json XdsAuditLoggerRegistry::ConvertXdsAuditLoggerConfig(
   }
   return Json::FromObject({{std::string(name), std::move(config)}});
 }
+
 }  // namespace grpc_core
